requirements.txt
streamlit
pandas
supabase
sodapy
requests
python-dotenv
altair
pydantic
google-generativeai